back_end/requirements.txt
Django==3.0
djangorestframework==3.11.0
psycopg2-binary>=2.8
django-debug-toolbar==2.2
django-extensions==3.0.0
Pillow==7.0.0
